bird_android_automation/requirements.txt
Appium-Python-Client>=4.0.0
selenium>=4.15.0
opencv-python-headless
numpy
